requirements.txt
# Core dependencies
google-adk==1.0.0
fastapi==0.115.0
uvicorn[standard]>=0.34.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
asyncpg==0.29.0

# ADK DatabaseSessionService dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Missing dependency for google-adk
deprecated>=1.2.14

# Microsoft Graph for Azure AD integration
msgraph-sdk>=1.0.0
azure-identity>=1.12.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Optional: For production logging
structlog==24.4.0
